refactor(proto): move proto message hub to its own module
Clean up the lib.rs file by moving the proto message hub to its own
module.
